backend/requirements.txt
Django==4.2.26
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
python-decouple==3.8